+++ src/Symfony/Component/Messenger/Bridge/Redis/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+6.0
+---
+
+ * Remove option `tls`
+ * Using invalid options will throw a `LogicException`
+
 5.3
 ---